smbinfo: update help text
Add description for fileallinfo query option.

Note that there are eight other recently added query options, but they
are mostly a subset a "fileallinfo" so could be of little value
(and may even be very confusing if we documented all nine in the
help text in smbinfo, instead of just this one).  The man page
has a full description of them.

Signed-off-by: Steve French <stfrench@microsoft.com>