Add missing position handling to mount parameters gid/backup_gid/snapshot
The code tries to optimise for the last parameter not needing to update
the position which means that every time a new one is added to the end
by copying and pasting, the string position is not updated.

That makes it impossible to use backup_uid=/backup_gid=/snapshot= after
gid= or snapshot= after backup_gid= because part of the string is
overwritten and contains invalid keys like "gbackup_uid".

Prepare for the next parameter to be added on the end by updating the
position for snapshot= even though it will be unused.